requirements.txt
streamlit==1.37.1
pandas==2.2.2
PyMuPDF==1.24.10
python-docx==1.1.2
reportlab==4.2.2
wordcloud==1.9.3
streamlit
pandas
PyPDF2
pymupdf
python-docx
nltk
scikit-learn
spacy
sumyspacy
sumy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
fastapi
uvicorn
pymongo
python-multipart
requests
beautifulsoup4
spacy
